test(ui): restructure Common classes, relocating methods to actions class from Elements class
